requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
pillow
django_extensions
pandas
django-cors-headers
djangorestframework